requirements.txt
anthropic>=0.40.0
faster-whisper>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
pydantic>=2.5.0
rich>=13.7.0
httpx>=0.26.0
pytz>=2024.1
